+++ src/pairwiselistmatrix.jl
@@ -527,7 +527,9 @@ end
                               bc::Base.Broadcast.Broadcasted{Nothing}) where {T, D, VT}
     axes(dest) == axes(bc) || Base.Broadcast.throwdm(axes(dest), axes(bc))
     bc_ = Base.Broadcast.preprocess(dest, bc)
-    @iterateupper dest true list[k] = :($bc_)[CartesianIndex(i,j)] # slow: bc_ has a plm
+    apply2upper(dest; use_diag=true) do list, k, i, j
+        list[k] = bc_[CartesianIndex(i,j)] # slow: bc_ has a plm
+    end
     return dest
 end
 
@@ -1036,11 +1038,11 @@ function to_table(plm::PairwiseListMatrix; diagonal::Bool = true, labels = getla
     N = plm.nelements
     table = Array{Any}(undef, diagonal ? div(N*(N+1),2) : div(N*(N-1),2), 3)
     t = 0
-    @iterateupper plm diagonal begin
-        :($t) += 1
-        :($table)[:($t), 1] = :($labels)[i]
-        :($table)[:($t), 2] = :($labels)[j]
-        :($table)[:($t), 3] = list[k]
+    apply2upper(plm; use_diag=diagonal) do list, k, i, j
+            t += 1
+            table[t, 1] = labels[i]
+            table[t, 2] = labels[j]
+            table[t, 3] = list[k]
     end
     table
 end
@@ -1092,14 +1094,15 @@ function to_dict(plm::PairwiseListMatrix{T,D,TV};
     J = Array{String}(undef, L)
     K = Array{T}(undef, L)
     t = 0
-    @iterateupper plm diagonal begin
-        :($t) += 1
-        :($I)[:($t)] = :($labels)[i]
-        :($J)[:($t)] = :($labels)[j]
-        :($K)[:($t)] = list[k]
+    apply2upper(plm; use_diag=diagonal) do list, k, i, j
+        t += 1
+        I[t] = labels[i]
+        J[t] = labels[j]
+        K[t] = list[k]
     end
     Dict(:i => I, :j => J, :values => K)
 end
+
 
 function to_dict(nplm::NamedArray{T,2,PairwiseListMatrix{T,D,TV},DN};
                  diagonal::Bool = true,
@@ -1214,9 +1217,12 @@ function DelimitedFiles.writedlm(filename::String,
                        delim::Char = '\t',
                        labels::Vector{String} = getlabels(plm)) where {T,D,TV}
     open(filename, "w") do fh
-        @iterateupper plm diagonal println(:($fh), :($labels)[i], :($delim), :($labels)[j], :($delim), list[k])
+        apply2upper(plm; use_diag=diagonal) do list, k, i, j
+            println(fh, labels[i], delim, labels[j], delim, list[k])
+        end
     end
 end
+
 
 function DelimitedFiles.writedlm(filename::String,
                        nplm::NamedArray{T,2,PairwiseListMatrix{T,D,TV},DN};